requirements.txt
# TeachClear — Python dependencies
# Install with: pip install -r requirements.txt

pyTelegramBotAPI
python-dotenv
google-genai

